backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic[email]==2.10.3
python-dotenv==1.0.1
hypothesis==6.122.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
google-generativeai==0.3.2
motor==3.6.0
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
